CLDR-17759 Move non-Info-Panel code from cldrInfo to cldrTable
-Move cldrInfo.listen and related code to cldrTable.listen since its effects are not limited to the Info Panel

-This PR is intended not to change behavior, strictly as refactoring in preparation for further changes

-Comments
